Deprecate ddtrace.encoding module (#2346)
* Deprecate ddtrace.encoding module

* isort fix
